Add sanity check for clipboard contents removal.

+++ mixer/src/main.C
@@ -346,10 +346,13 @@ main ( int argc, char **argv )
     main_window = NULL;
 
     /* Delete clipboard contents because if the strip contains custom data then it will accumulate */
-    std::string remove_clipboard = "exec rm -r ";
-    remove_clipboard += clipboard_dir;
-    remove_clipboard += "/*";
-    system(remove_clipboard.c_str());
+    if(clipboard_dir)
+    {
+        std::string remove_clipboard = "exec rm -r ";
+        remove_clipboard += clipboard_dir;
+        remove_clipboard += "/*";
+        system(remove_clipboard.c_str());
+    }
 
     MESSAGE( "Your fun is over" );
 }
